fix: вынесение сигналов в отдельный файл
