+++ articles/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class ArticlesConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'articles'
+
+    def ready(self):
+        import articles.signals  # noqa
